Move functions to helpers/, as it will enable unit testing
As Pascal mentioned,
this will be a better place for the functions.
Luckily, only isVideoSizeExcessive() needs to
be exported from helpers/.
